backend/requirements.txt
Flask==3.1.0
Flask-CORS==6.0.1
python-dotenv==1.1.1
mysql-connector-python==8.3.0
easyocr==1.7.1
Pillow==10.3.0
pdf2image==1.17.0
numpy==1.26.4
bcrypt==4.2.0
PyJWT==2.9.0